refac: move render methods from MethodInvokerPage to other class
